requirements.txt
# Core dependencies for e-Stat API extraction
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# BigQuery client (for loading data to BigQuery)
google-cloud-bigquery>=3.10.0

# Optional: Data validation
pydantic>=2.0.0
